requirements.txt
python-telegram-bot==20.3
peewee==3.16.2
python-dotenv==1.0.0